move _order.py into _encryption.py
